Deployed: real cell data live from garage banks 1-3; O-1 decode pinned
Node A deployed by the BMS rack; everything after the one bootstrap
power-cycle was done remotely over the new push-OTA + MQTT ops path.

- components/ota: push-model OTA receiver (POST /ota :3765), dual-slot
  4MB partitions + bootloader rollback; image confirmed only once WiFi
  is up, so a bad build self-reverts. tools/ota_push.py = host updater.
  ota_start now unconditional (late WiFi join still exposes the port).
- MQTT ops: bridge/cmd/{reboot,scan,rawcap,nvsclear}; app writes are
  logged to <id>/appwrite; raw notify capture to <id>/raw. These four
  are what made remote diagnosis possible.
- Real units: CFG_BMS = BMS_0-00 / 'BMS 1-01' / 'BMS 2-02' / BMS_3-03
  (verified via scan dump), CFG_NUM_UNITS=4, unit 0 parked (flaps;
  Telink OUI, different HW rev). nvsclear killed the stale bench
  harvest that made the O-9 layout check exclude all real units.
- jk_proto: decode offsets pinned from live JK-PB2A16S20P fw 19.31
  captures and cross-checked against the .90 aggregator (current
  -0.133 A exact). Cell mask @70 drives cell_count (16S). Settings and
  device-info decode filled in. synth_frames in lockstep; host test
  gains a real-capture regression (26/26).
- Streaming trigger: fw 19.31 starts the 0x02 stream only after 0x97
  THEN 0x96 in one session; bootstrap pair on every link-up, 0x97
  nonce-filled like the official app.

+++ components/jk_proto/include/jk_proto.h
@@ -61,9 +61,9 @@ typedef struct {
     int32_t   current_ma;        /* signed; + = charge                        */
     int32_t   power_mw;
     uint8_t   soc_pct;
-    int16_t   temp_dc[4];        /* deci-°C; VERIFY scale/positions            */
+    int16_t   temp_dc[4];        /* deci-°C: [0]=T1 [1]=T2 [2]=power tube      */
     uint8_t   temp_count;
-    uint16_t  cycle_count;
+    uint32_t  cycle_count;       /* u32 @182 (bench-verified)                  */
     bool      balancing;
     bool      charging;
     bool      discharging;
